Drop aiida-core dependency
