+++ setup.cfg
@@ -23,7 +23,6 @@ install_requires =
     CacheControl~=0.12
     Jinja2>=2.11.3,<4
     MarkupSafe==2.0.1
-    aiida-core>=1.0.0,<3
     cachetools~=4.1
     click>=7.0,<9
     click-spinner~=0.1
