+++ tests/conftest.py
@@ -26,3 +26,23 @@ def generate_app():
         return app
 
     return _generate_app
+
+
+_MONKEYPATCHED_INSTALLED_PACKAGES = [
+    {"name": "aiida-core", "version": "2.2.1"},
+    {"name": "jupyter_client", "version": "7.3.5"},
+]
+
+
+@pytest.fixture
+def installed_packages(monkeypatch):
+    """change the return of pip_list.
+    This is to mimic the pip list command output, which returns a json string represent
+    the list of installed packages."""
+    from aiidalab.utils import FIND_INSTALLED_PACKAGES_CACHE
+
+    FIND_INSTALLED_PACKAGES_CACHE.clear()  # clear the cache
+    monkeypatch.setattr(
+        "aiidalab.utils._pip_list",
+        lambda _: _MONKEYPATCHED_INSTALLED_PACKAGES,
+    )
